+++ src/deep/merge.clj
@@ -1,72 +1,51 @@
 (ns deep.merge
   (:gen-class))
 
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; These are variables needed for the main 'deep-merge' method of this project ;;
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(defonce ^:dynamic *vector-merge-method* nil)
-(defonce ^:dynamic *merge-mixed-vector-map-sets* false)
+(defn- deep-coll-merge-with
+  "The base method used for recursive collection merging"
+  [collection-merge-method non-collection-merge-method & vals]
+  (let [this-method (partial deep-coll-merge-with collection-merge-method non-collection-merge-method)]
+    (cond
+      (every? map? vals) (apply merge-with this-method vals)
+      (every? coll? vals) (apply collection-merge-method vals)
+      :else (apply non-collection-merge-method vals))))
 
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; This is the main method that can be used by redefining the above variables ;;
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(defn deep-merge
-  "Based on the various environment variables in this namespace, merge two maps or vectors.
-  If a merge of the objects cannot happen, the last object is returned."
-  ([] nil)
-  ([& obj-args]
-   (cond
-     (every? map? obj-args) (apply merge-with deep-merge obj-args)
-     (every? vector? obj-args) (apply *vector-merge-method* obj-args)
-     (and *merge-mixed-vector-map-sets* (every? #(or (map? %) (vector? %)) obj-args)) (deep-merge (map #(if (map? %) [%] %)))
-     :always (last obj-args))))
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Collection Merge Methods ;;
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+(defn- concat-coll-merge
+  "Puts all items from all args blindly into one vector"
+  [& args]
+  (->> args
+       (apply concat)
+       (into [])))
+(defn- distinct-concat-coll-merge
+  "Puts all items from all args into one deduplicated vector"
+  [& args]
+  (->> args
+       (apply concat)
+       (distinct)
+       (into [])))
+(defn- index-coll-merge
+  "Looks at each index of all args and runs deep-coll-merge-with on all items at each index"
+  [non-collection-merge-method & args]
+  (let [max-index (apply max (map count args))
+        values-at-index (fn [index] (->> args
+                                         (filter #(< index (count %)))
+                                         (map #(nth % index))))
+        merge-index #(apply deep-coll-merge-with index-coll-merge non-collection-merge-method (values-at-index %))]
+    (->> (range 0 max-index)
+         (map merge-index)
+         (into []))))
 
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; These methods define the various methods by which vectors can be merged ;;
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(defmacro with-vector-merge-method
-  "Adds a vector merge strategy to deep-merge."
-  [method & body]
-  `(binding [*vector-merge-method* ~method]
-     ~@body))
-(defn vector-blind-merge
-  "Blindly puts all vectors passed in into one vector.
-  If not all items are vectors, the last item is returned"
-  [& vectors]
-  (if (every? vector? vectors)
-    (apply concat vectors)
-    (last vectors)))
-(defn vector-index-merge
-  "Looks at each index across all vectors and runs deep-merge on that.
-  If not all items are vectors, the last item is returned."
-  [& vectors]
-  (if (every? vector? vectors)
-    (let [max-index (apply max (map count vectors))
-          get-all-indexes (fn [index] (filter
-                                        #(not (nil? %))
-                                        (map #(nth % index nil) vectors)))]
-      (for [i (range 0 max-index)] (apply deep-merge (get-all-indexes i))))
-    (last vectors)))
-(defn vector-blind-merge-with-dedupe
-  "Does the same thing as blind merge, but applies distinct to the resulting vector.
-  If not all items are vectors, the last item is returned."
-  [& vectors]
-  (if (every? vector? vectors)
-    (distinct (apply concat vectors))
-    (last vectors)))
-
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; Additional options for 'deep-merge' ;;
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(defmacro with-mixed-vector-map-merging
-  "deep-merges within this macro will allow mixed sets of vectors/maps to be merged.
-  This happens by taking maps and putting them into the first index of new vectors.
-  After the conversion happens, deep merge is called again."
-  [& body]
-  `(binding [*merge-mixed-vector-map-sets* true]
-     ~@body))
-
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; Setting defaults for 'deep-merge' ;;
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(alter-var-root (var *vector-merge-method*) (constantly vector-blind-merge))
+;;;;;;;;;;;;;;;;;;;;;
+;; Wrapper Methods ;;
+;;;;;;;;;;;;;;;;;;;;;
+(def concat-merge-with (partial deep-coll-merge-with concat-coll-merge))
+(def concat-merge (partial concat-merge-with #(last %&)))
+(def distinct-merge-with (partial deep-coll-merge-with distinct-concat-coll-merge))
+(def distinct-merge (partial distinct-merge-with #(last %&)))
+(defn- index-merge-with
+  [non-collection-merge-method & vals]
+  (apply deep-coll-merge-with (partial index-coll-merge non-collection-merge-method) non-collection-merge-method vals))
+(def index-merge (partial index-merge-with #(last %&)))